complete builder.py refactor: replace midas.py with midas2.py (same with test_midas.py)
History for midas2.py:
  commit 3a01d7fe1afa6ec34d8869f477f5c24e04250b1a
  Author: Ray Plante <raymond.plante@nist.gov>
  Date:   Thu Nov 29 09:23:29 2018 -0500

      builder2, midas2: tweaked messaging

  commit f711f901268ec4825383bccd72c977dabf6a5bf8
  Author: Ray Plante <raymond.plante@nist.gov>
  Date:   Wed Nov 28 22:08:34 2018 -0500

      builder refactor: complete integration: have modules use midas2;
                        debug; update tests for changes to test data

  commit 2ce331136ce64edffab277fe3a59f4a6e61e1fc7
  Author: Ray Plante <raymond.plante@nist.gov>
  Date:   Wed Nov 28 22:06:55 2018 -0500

      midas2.py: reduce the number of calls to nerdm_record()

  commit 268c8e8e285aaa2280eed1161401c64a61cf7790
  Author: Ray Plante <raymond.plante@nist.gov>
  Date:   Wed Nov 28 14:23:45 2018 -0500

      builder refactor: debug midas2

  commit b1e32fb83e539cd62227395f05db45408783145d
  Author: Ray Plante <raymond.plante@nist.gov>
  Date:   Wed Nov 28 05:48:46 2018 -0500

      more integration of builder2 into midas2 (still need to test updates)

  commit c8250080d39bc32e86821e450da460e743e34219
  Author: Ray Plante <raymond.plante@nist.gov>
  Date:   Tue Nov 27 09:56:51 2018 -0500

      start integration of builder2.py into midas via midas2.py
